+++ src/features/edit-exam/types.ts
@@ -1,0 +1,5 @@
+export interface GradeFormData {
+  score: number;
+  weight: number;
+  comment: string;
+}
